another memory optimisation
This time memory leaked if lots of UUIDs needed to be pretty-printed, as in
a get or drop of many files. Essentially the same strict read buffering
problem that affected the LocationLog underneath though.

uuidMap really could stand to be cached, as the uuid log is read many times
in this case. But it is a fairly edge case.