bugfix
The object's directory might exist if the backend picked the same key
as was already present. That could happen, for example, if the sha1 is the
same.

Note that I chose to go ahead and replace the old content with the new.
We don't know if they are the same (even with sha1, their timestamp or
perms could differ), so have to assume the newer one is preffered. Although
it won't propigate to other annexes, so it had better not be significantly
different!