remove bad information
You do not need to pull from a remote to make git-annex aware of it.
Pulling only makes git know about remote/master branches, but git-annex
does not care about that. When git-annex encounters a remote in .git/config
that has no recorded annex.uuid, it connects to it to get the uuid
automatically, and will immediately start using it.

A later part of this walkthrough shows how to pull from a remote
in order to get its location tracking info, so that git-annex can retrieve
files from it. That's the right place to document the need to pull from
a remote periodically.